\name{dat.employee}
\docType{data}
\alias{dat.employee}
\title{Data: Employees}
\description{Some human resource data on 37 employees with 6 variables. The following is the contents of the corresponding file of variable labels that would be read from an external \code{csv} file with \code{\link{rad.labels}}.\cr\cr
Years,"Years Employed in the Company"\cr
Gender,"Male or Female"\cr
Dept,"Department Employed"\cr
Salary,"Annual Salary (USD)"\cr
Satisfaction,"Satisfaction with Work Environment"\cr
HealthPlan,"1=GoodHealth, 2=YellowCross, 3=BestCare"\cr
}
\usage{data(dat.employee)}
\format{A data table with 37 observations.}
\source{author}
\keyword{datasets}

