\name{dat.twogroup}
\docType{data}
\alias{dat.twogroup}
\title{Data: ShipTimes of Two Suppliers}
\description{One grouping variable that specifies the supplier, Supplier, and one response variable, ShipTime. There are two suppliers, S1 and S2.}
\usage{data(dat.twogroup)}
\format{A data table with 31 observations.}
\source{author}
\keyword{datasets}

