\name{dataAnova_1way}
\docType{data}
\alias{dataAnova_1way}
\title{Data: Anova_1way}

\description{To study the impact of arousal on the ability to complete a task, 24 laboratory rats were randomly and equally divided into three groups of eight, and then given one of three dosages of an arousal inducing drug: 0, 5, and 10 milligrams. Following the dosage, each rat completed a maze to obtain a food reward. The response (dependent) variable is the Time in seconds to complete the maze.}

\usage{data(dataAnova_1way)}

\format{A data table with 24 rows of data and 2 columns, with variables Dosage and Time.}

\source{author}
\keyword{datasets}
