\name{dataAnova_2way}
\docType{data}
\alias{dataAnova_2way}
\title{Data: Anova_2way}

\description{Laboratory rats were randomly and equally divided into groups, and then given one of three dosages of an arousal inducing drug: 0, 5, and 10 milligrams. Following the dosage, each rat completed either an easy or a hard maze to obtain a food reward. The response (dependent) variable is the Time in seconds to complete the maze.}

\usage{data(dataAnova_2way)}

\format{A data table with 48 rows of data and 3 columns.}

\source{author}
\keyword{datasets}
