\name{dataBodyMeas}
\docType{data}
\alias{dataBodyMeas}
\title{Data: Body Measurements}

\description{Body measurements of 170 women and 170 men who purchased motorcycle clothing from Gerbing's Heated Clothing, Inc., now Gordon's Heated Clothing, LLC.
}

\usage{data(dataBodyMeas)}

\format{A data table with 340 observations and the following 7 variables.

Gender, "M" or "F" (factor)\cr
Weight (integer in pounds)\cr
Height (integer in inches)\cr
Waist (integer in inches)\cr
Hips (integer in inches)\cr
Chest (integer in inches)\cr
Hand (numeric, circumference of hand in inches to nearest quarter of an inch)\cr
Shoe (numeric, size including half sizes)
}

\source{author}
\keyword{datasets}
