\name{dataCars93}
\docType{data}
\alias{dataCars93}
\title{Data: Cars93}

\description{1993 New Car Data.}

\usage{data(dataCars93)}
\format{A data table with 93 observations.\cr

Variables\cr
Make, Model, Type, MinPrice, MidPrice, MaxPrice, MPGcity, MPGhiway, Airbags, DriveTrain, Cylinders, Engine, HP, RPM, RevMile, Manual, FuelCap, PassCap, Length, Wheelbase, Width, Uturn, RearSeat, LugCap, Weight, Source
}

\source{Lock, R. H. (1993). 1993 new car data. Journal of Statistics Education, 1(1).}
\keyword{datasets}
