\name{dataEmployee}
\docType{data}
\alias{dataEmployee}
\title{Data: Employees}

\description{Some human resource data on 37 employees with 6 variables. Variable labels are included in the data file.}

\usage{data(dataEmployee)}
\format{A data table with 37 observations.

Years,"Years Employed in the Company"\cr
Gender,"Male or Female"\cr
Dept,"Department Employed"\cr
Salary,"Annual Salary (USD)"\cr
Satisfaction,"Satisfaction with Work Environment"\cr
HealthPlan,"1=GoodHealth, 2=YellowCross, 3=BestCare"
}

\source{author}
\keyword{datasets}
