\name{dataFreqTable99}
\docType{data}
\alias{dataFreqTable99}
\title{Data: Joint Frequency Table}

\description{
Based on a survey of university students, the joint frequencies for two variables are reported. One variable is Race and the other is undergraduate Class.

\code{
 Level Asian Latino Black White
    FR    33     58     6   105
    SO    41     79     9   207
    JR    86    179    27   484
    SR   143    214    31   824
}

The data file consists just of the frequencies, the numbers, without any labels.
}

\usage{data(dataFreqTable99)}
\format{A table of joint frequencies or Race and Level.

Race: Asian, Latino, Black, White\cr
Class: FR, SO, JR, SR\cr
}

\source{author}
\keyword{datasets}


