\name{dataGroups2}
\docType{data}
\alias{dataGroups2}
\title{Data: ShipTimes of Two Suppliers}

\description{Completely Randomized design, one-factor with two levels (CR-2): One grouping variable that specifies the supplier, Supplier, and one response variable, ShipTime. There are two suppliers, S1 and S2.}

\usage{data(dataGroups2)}

\format{A data table with 31 observations.}

\source{author}
\keyword{datasets}
