\name{dataJackets}
\docType{data}
\alias{dataJackets}
\title{Data: Motorcycle Type and Thickness of Jacket}

\description{Two variables, one is type of motorcycle and the other is the thickness of the purchased jacket.}

\usage{data(dataJackets)}
\format{A data table with 1025 observations.

Bike,"Type of Motorcycle, Honda or BMW"\cr
Jacket,"Lite, Med or Thick"\cr
}

\source{author}
\keyword{datasets}


