\name{dataLearn}
\docType{data}
\alias{dataLearn}
\title{Data: Distributed vs Massed Practice}

\description{Completely Randomized design, one-factor with two levels (CR-2): One grouping variable that specifies type of learning, distributed or massed practice, and one response variable, Learning.}

\usage{data(dataLearn)}

\format{A data table with 34 observations.}

\source{author}
\keyword{datasets}
