\name{dataMach4}
\docType{data}
\alias{dataMach4}
\title{Data: Machiavellianism Data}
\description{Likert data responses to Christie and Giess's Mach4 scale.

All Likert items assessed on a 6-point scale from 0: Strongly Disagree to 5: Strongly Agree.
}

\usage{data(dataMach4)}

\format{A data table with 351 observations.

Gender, 1 column, 0:Male, 1:Female\cr
Mach IV, 20 Likert items, named m01, m02, \ldots, m20 \cr

Christie and Geiss Mach IV Scale\cr
-------------------------------------------\cr
1. Never tell anyone the real reason you did something unless it is useful to do so.\cr
2. The best way to handle people is to tell them what they want to hear.\cr
3. One should take action only when sure it is morally right.\cr
4. Most people are basically good and kind.\cr
5. It is safest to assume that all people have a vicious streak and it will come out when they are given a chance.\cr
6. Honesty is the best policy in all cases.\cr
7. There is no excuse for lying to someone else.\cr
8. Generally speaking, people won't work hard unless they're forced to do so.\cr
9. All in all, it is better to be humble and honest than to be important and dishonest.\cr
10. When you ask someone to do something for you, it is best to give the real reasons for wanting it rather than giving reasons which carry more weight.\cr
11. Most people who get ahead in the world lead clean, moral lives.\cr
12. Anyone who completely trusts anyone else is asking for trouble.\cr
13. The biggest difference between most criminals and other people is that the criminals are stupid enough to get caught.\cr
14. Most people are brave.\cr
15. It is wise to flatter important people.\cr
16. It is possible to be good in all respects.\cr
17. Barnum was wrong when he said that there's a sucker born every minute.\cr
18. It is hard to get ahead without cutting corners here and there.\cr
19. People suffering from incurable diseases should have the choice of being put painlessly to death.\cr
20. Most people forget more easily the death of a parent than the loss of their property.
}

\source{author}

\references{
Christie, R., & Geis, F. L., (1970).  Studies in Machiavellianism.  New York:  Academic Press.

Hunter, J. E., Gerbing, D. W., and Boster, F. J. (1982). Machiavellian beliefs and personality: The construct invalidity of the Machiavellian dimension. Journal of Personality and Social Psychology, 43, 1293-1305.
}

\keyword{datasets}

