\name{dataReading}
\docType{data}
\alias{dataReading}
\title{Data: Reading Ability}

\description{Reading ability test score and also verbal aptitude test score, number of absences from school and family income in USD $1000's.}

\usage{data(dataReading)}

\format{A data table with 100 observations.}

\source{author}
\keyword{datasets}
