\name{dataStockPrice}
\docType{data}
\alias{dataStockPrice}
\title{Data: Stock price of Apple, IBM and Intel from 1985 through June of 2022}

\description{Monthly adjusted stock price of Apple, IBM and Intel from 1985 through June of 2022 from finance.yahoo.com.}

\usage{data(dataStockPrice)}

\format{A data table in long format with three variables: month, Company and Price. The variable month is a date expression expressed as a four digit year, followed by a month, then a day, separated by either dashes. A total of 1350 rows, 450 rows per company.}

\source{author}
\keyword{datasets}
