\name{dataWeightLoss}
\docType{data}
\alias{dataWeightLoss}
\title{Data: WeightLoss}

\description{The weights of 10 people were recorded. Then they entered a weight loss program. Following completion of the program, their weights were once again recorded. Data are simulated.}

\usage{data(dataWeightLoss)}

\format{A data table with 10 observations.}

\source{author}
\keyword{datasets}
