\name{Help}
\alias{Help}

\title{Help System for Statistics by Topic that Suggests Related Functions}

\description{

The lessR \code{(Help()} function was a kind of mini-vignette. Now that R provides these, the \code{Help()} function is removed, replaced by the more extensive and informative vignettes. For a list of all vigenettes, enter \code{vignette(package=\"lessR\")}.
}

\usage{
Help(topic=NULL)
}


\arguments{
  \item{topic}{... Removed ...}
}
