\name{sort_by}
\alias{sort_by}

\title{Sort_by the Rows of a Data Frame}

\description{
The R folks named their own function  sort_by()  years after lessR used the name. To avoid confusion, the lessR function is now renamed:  order_by()

The name  sort_by is deprecated, to be removed in the future.
}

\usage{
sort_by
}

