\name{xU}
\alias{xU}

\title{Text Processing: Capitalize First Letter of a Word}

\description{Capitalize the first letter of a word. Primarily for internal use in text processing of \code{knitr} output. Not usually referenced by the user.
}

\usage{
xU(x)
}


\arguments{
  \item{x}{The character string (word) for which to capitalize the first letter.}
}


\details{
Input is a single word.  Output is the word with its first letter capitalized.
}

\author{David W. Gerbing (Portland State University; \email{gerbing@pdx.edu})}

\examples{
xU("the")
}

