\name{dataAnova_rb}
\docType{data}
\alias{dataAnova_rb}
\title{Data: Anova_rb}

\description{Eight people, with differing amounts of muscle strength, took one of four different pre-workout supplements and then did a bench press of 125 lbs as many times as possible. Each person did four workouts at four different times, with a different supplement before each workout. The experimenter randomized the presentation order of the supplements to each person to avoid any artifacts from presentation order.}

\usage{data(dataAnova_rb)}

\format{A data table with 7 rows of data and 5 columns.}

\source{author}
\keyword{datasets}
