\name{dataCars93}
\docType{data}
\alias{dataCars93}
\title{Data: Cars93}

\description{1993 New Car Data.}

\usage{data(dataCars93)}
\format{A data table with 93 observations and 25 variables.\cr

Variables\cr
-------------------------------------------\cr
Make: Model\cr
Type: Small, Sporty, Compact, Midsize, Large, Van\cr
MinPrice: Minimum Price (in $1,000) - Price for basic version of this model\cr
MidPrice: Midrange Price (in $1,000) - Average of Min and Max prices\cr
MaxPrice: Maximum Price (in $1,000) - Price for a premium version\cr
MPGcity: City MPG\cr
MPGhiway: Highway MPG\cr
Airbags: 0 = none, 1 = driver only, 2 = driver & passenger\cr
DriveTrain: 0 = rear wheel drive, 1 = front wheel drive, 2 = all wheel drive\cr
Cylinders: Number of cylinders\cr
Engine: Engine size (liters)\cr
HP: maximum Horsepower\cr
RPM: revolutions per minute at maximum horsepower\cr
RevMile: Engine revolutions per mile in highest gear\cr
Manual: Manual transmission available, 0 = No, 1 = Yes\cr
FuelCap: Fuel tank capacity (gallons)\cr
PassCap: Passenger capacity (persons)\cr
Length: Length (inches)\cr
Wheelbase: Wheelbase (inches)\cr
Width: Width (inches)\cr
Uturn: U-turn space (feet)\cr
RearSeat: Rear seat room (inches)\cr
LugCap: Luggage capacity (cu. ft.)\cr
Weight: Weight (pounds)\cr
Source: 0=non-USA manufacturer, 1=USA manufacturer\cr
}

\source{Lock, R. H. (1993). 1993 new car data. Journal of Statistics Education, 1(1).}
\keyword{datasets}
