\name{dataEmployee}
\docType{data}
\alias{dataEmployee}
\title{Data: Employees}

\description{Some human resource data on 37 employees with 6 variables. Variable labels and variable units are included in the data file.}

\usage{data(dataEmployee)}
\format{A data table with 37 observations.

Years,"Years Employed in the Company"\cr
Gender,"Male or Female"\cr
Dept,"Department Employed"\cr
Salary,"Annual Salary (USD)"\cr
JobSat,"JobSat with Work Environment"\cr
Plan,"1=GoodHealth, 2=YellowCross, 3=BestCare"\cr
Pre,"Test score on legal issues before instruction"\cr
Post,"Test score on legal issues after instruction"
}

\source{author}
\keyword{datasets}
