\name{dataEmployee_lbl}
\docType{data}
\alias{dataEmployee_lbl}
\title{VariableLabels: Employee Data Set}

\description{For the data on 37 employees with 6 variablesc, includes the variable labels and variable units.}

\usage{data(dataEmployee_lbl)}
\format{Variable labels.

Years,"Years Employed in the Company"\cr
Gender,"Male or Female"\cr
Dept,"Department Employed"\cr
Salary,"Annual Salary (USD)"\cr
Satisfaction,"Satisfaction with Work Environment"\cr
HealthPlan,"1=GoodHealth, 2=YellowCross, 3=BestCare"\cr
Pre,"Test score on legal issues before instruction"\cr
Post,"Test score on legal issues after instruction"
}

\source{author}
\keyword{datasets}
