\name{dataStockPrice}
\docType{data}
\alias{dataStockPrice}
\title{Data: Stock price of Apple, IBM and Intel from 1980 through 2017}

\description{Monthly adjusted stock price of Apple, IBM and Intel from 1980 through 2017 from finance.yahoo.com.}

\usage{data(dataStockPrice)}

\format{A data table in tidy format with three variables: month, Company and Price. The variable month is a date expression expressed as a four digit year, followed by a month, then a day, separated by either dashes. A total of 1308 rows, 436 rows per company.}

\source{author}
\keyword{datasets}
