\name{dataStockPrice}
\docType{data}
\alias{dataStockPrice}
\title{Data: Stock price of Apple, IBM and Intel from 1985 through May of 2024}

\description{Monthly adjusted stock price of Apple, IBM and Intel from 1985 through May of 2024 from finance.yahoo.com.}

\usage{data(dataStockPrice)}

\format{A data table in long format with four variables: Month, Company, Price, and Volume. The variable Month is a date expression expressed in the ISO standard as a four-digit year, followed by the two-digit month, then the two-digit day, separated by dashes. A total of 1419 rows, 473 rows per company.}

\source{author}
\keyword{datasets}
